thread fractions through STV and the things it touches

+++ src/votekit/elections/transfers.py
@@ -4,12 +4,12 @@ from typing import Optional, Union
 
 from votekit.ballot import RankBallot
 from votekit.pref_profile import RankProfile
-from votekit.types import Candidate
+from votekit.types import Candidate, Numeric
 
 
 def fractional_transfer(
     winner: Candidate,
-    fpv: float,
+    fpv: Numeric,
     ballots: Union[tuple[RankBallot], list[RankBallot]],
     threshold: int,
 ) -> tuple[RankBallot, ...]:
@@ -19,7 +19,7 @@ def fractional_transfer(
     Args:
         winner (Candidate): Candidate to transfer votes from.
             Candidate can be a string or integer.
-        fpv (float): Number of first place votes for winning candidate.
+        fpv (Numeric): Number of first place votes for winning candidate.
         ballots (Union[tuple[RankBallot], list[RankBallot]]): List of Ballot objects.
         threshold (int): Value required to be elected, used to calculate transfer value.
 
@@ -56,7 +56,7 @@ def fractional_transfer(
 
 def random_transfer(
     winner: Candidate,
-    fpv: float,
+    fpv: Numeric,
     ballots: Union[tuple[RankBallot], list[RankBallot]],
     threshold: int,
     *,
@@ -69,7 +69,7 @@ def random_transfer(
     Args:
         winner (Candidate): Candidate to transfer votes from.
             Candidates can be strings, integers, or mix of both.
-        fpv (float): Number of first place votes for winning candidate.
+        fpv (Numeric): Number of first place votes for winning candidate.
         ballots (Union[tuple[RankBallot], list[RankBallot]]): List of Ballot objects.
         threshold (int): Value required to be elected, used to calculate transfer value.
         rng (random.Random, optional): Standard library random number generator. Pass a seeded
